move coke::make_task from coke.h to make_task.h
