+++ include/coke/coke.h
@@ -1,5 +1,5 @@
-#ifndef COKE_BASIC_H
-#define COKE_BASIC_H
+#ifndef COKE_COKE_H
+#define COKE_COKE_H
 
 #include <functional>
 #include <concepts>
@@ -21,34 +21,7 @@
 #include "coke/semaphore.h"
 #include "coke/mutex.h"
 #include "coke/shared_mutex.h"
+#include "coke/future.h"
+#include "coke/make_task.h"
 
-namespace coke {
-
-template<typename FUNC, typename... ARGS>
-    requires std::invocable<FUNC, ARGS...>
-auto make_task(FUNC &&func, ARGS&&... args) {
-    using result_t = std::invoke_result_t<FUNC, ARGS...>;
-    static_assert(is_task_v<result_t>, "FUNC(ARGS...) must return coke::Task<T>");
-
-    result_t res;
-    if constexpr (std::is_rvalue_reference_v<FUNC&&>) {
-        std::shared_ptr<FUNC> ctx = std::make_shared<FUNC>(std::forward<FUNC>(func));
-        res = std::invoke(*ctx, std::forward<ARGS>(args)...);
-        res.set_context(ctx);
-    }
-    else {
-        res = std::invoke(std::forward<FUNC>(func), std::forward<ARGS>(args)...);
-    }
-
-    return res;
-}
-
-template<typename FUNC, typename... ARGS>
-    requires std::invocable<FUNC, ARGS...>
-auto sync_call(FUNC &&func, ARGS&&... args) {
-    return sync_wait(make_task(std::forward<FUNC>(func), std::forward<ARGS>(args)...));
-}
-
-} // namespace coke
-
-#endif // COKE_BASIC_H
+#endif // COKE_COKE_H
